✨ feat: Add player_name parameter to create and join functions in IHost trait
This commit adds a new `player_name` parameter to the `create` and `join` functions in the `IHost` trait. Previously, these functions only accepted the `world` and `game_id` parameters. With this change, the functions now also accept a `player_name` parameter of type `felt252`. The `create` function creates a new host instance with the provided `world`, while the `join` function allows a player to join a specific game identified by `game_id`. The `player_name` parameter is used to set the name of the player participating in the game.